+++ src/__tests__/data/match-path-data.ts
@@ -1,5 +1,5 @@
 import { join, dirname } from "path";
-import { removeExtension } from "../../src/filesystem";
+import { removeExtension } from "../../filesystem";
 
 export interface OneTest {
   readonly name: string;
@@ -11,10 +11,18 @@ export interface OneTest {
   readonly addMatchAll?: boolean;
   readonly existingFiles: ReadonlyArray<string>;
   readonly requestedModule: string;
-  readonly extensions?: ReadonlyArray<string>;
+  readonly extensions: ReadonlyArray<string>;
   readonly packageJson?: {};
   readonly expectedPath: string | undefined;
 }
+
+const defaultExtensionsWhenRunningInTsNode = [
+  ".js",
+  ".json",
+  ".node",
+  ".ts",
+  ".tsx",
+];
 
 export const tests: ReadonlyArray<OneTest> = [
   {
@@ -26,6 +34,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "index.ts")],
     requestedModule: "lib/mylib",
     expectedPath: dirname(join("/root", "location", "mylib", "index.ts")),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should resolve to correct path when many are specified",
@@ -52,6 +61,7 @@ export const tests: ReadonlyArray<OneTest> = [
     ],
     requestedModule: "lib/mylib",
     expectedPath: dirname(join("/root", "location", "mylib", "index.ts")),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should locate path that matches with star and exists with extension",
@@ -69,6 +79,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "test.jpg")],
     requestedModule: "lib/test.jpg",
     expectedPath: join("/root", "location", "test.jpg"),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should locate path that matches without star and exists",
@@ -79,6 +90,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "foo.ts")],
     requestedModule: "lib/foo",
     expectedPath: removeExtension(join("/root", "location", "foo.ts")),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should resolve to parent folder when filename is in subfolder",
@@ -87,6 +99,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "index.ts")],
     requestedModule: "lib/mylib",
     expectedPath: dirname(join("/root", "location", "mylib", "index.ts")),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should resolve from main field in package.json",
@@ -95,9 +108,8 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "kalle.ts")],
     packageJson: { main: "./kalle.ts" },
     requestedModule: "lib/mylib",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylib", "kalle.ts")
-    ),
+    expectedPath: join("/root", "location", "mylib", "kalle.ts"),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should resolve from main field in package.json (js)",
@@ -107,22 +119,7 @@ export const tests: ReadonlyArray<OneTest> = [
     packageJson: { main: "./kalle.js" },
     requestedModule: "lib/mylib.js",
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylib.js", "kalle.js")
-    ),
-  },
-  {
-    name:
-      "should resolve from main field in package.json and correctly remove file extension",
-    absoluteBaseUrl: "/root/",
-    paths: { "lib/*": ["location/*"] },
-    existingFiles: [join("/root", "location", "mylibjs", "kalle.js")],
-    packageJson: { main: "./kalle.js" },
-    extensions: [".ts", ".js"],
-    requestedModule: "lib/mylibjs",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylib.js", "kalle.js"),
   },
   {
     name: "should resolve from list of fields by priority in package.json",
@@ -136,9 +133,7 @@ export const tests: ReadonlyArray<OneTest> = [
     ],
     extensions: [".ts", ".js"],
     requestedModule: "lib/mylibjs",
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "browser.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "browser.js"),
   },
   {
     name: "should ignore field mappings to missing files in package.json",
@@ -152,9 +147,7 @@ export const tests: ReadonlyArray<OneTest> = [
       browser: "./nope.js",
     },
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "kalle.js"),
   },
   {
     name: "should ignore advanced field mappings in package.json",
@@ -170,9 +163,7 @@ export const tests: ReadonlyArray<OneTest> = [
       browser: { mylibjs: "./browser.js", "./kalle.js": "./browser.js" },
     },
     extensions: [".ts", ".js"],
-    expectedPath: removeExtension(
-      join("/root", "location", "mylibjs", "kalle.js")
-    ),
+    expectedPath: join("/root", "location", "mylibjs", "kalle.js"),
   },
   {
     name: "should resolve to with the help of baseUrl when not explicitly set",
@@ -181,6 +172,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "mylib", "index.ts")],
     requestedModule: "mylib",
     expectedPath: dirname(join("/root", "mylib", "index.ts")),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should not resolve with the help of baseUrl when asked not to",
@@ -190,6 +182,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "mylib", "index.ts")],
     requestedModule: "mylib",
     expectedPath: undefined,
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should not locate path that does not match",
@@ -198,6 +191,7 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("root", "location", "mylib")],
     requestedModule: "mylib",
     expectedPath: undefined,
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
   {
     name: "should not resolve typings file (index.d.ts)",
@@ -208,5 +202,18 @@ export const tests: ReadonlyArray<OneTest> = [
     existingFiles: [join("/root", "location", "mylib", "index.d.ts")],
     requestedModule: "lib/mylib",
     expectedPath: undefined,
+    extensions: defaultExtensionsWhenRunningInTsNode,
+  },
+  {
+    name: "should resolve main file with cjs file extension",
+    absoluteBaseUrl: "/root/",
+    paths: {},
+    existingFiles: [join("/root", "mylib", "index.cjs")],
+    packageJson: {
+      main: "./index.cjs",
+    },
+    requestedModule: "mylib",
+    expectedPath: join("/root", "mylib", "index.cjs"),
+    extensions: defaultExtensionsWhenRunningInTsNode,
   },
 ];